Basic support for automatic copy updates
This is a totally breaking PR. It had to come sooner or later...

Before this patch, copier had some ugly problems:

- If a file was called `--exclude`, you couldn't add it to `--extra-paths` due to using `nargs='*'` in an option. It is a weird thing that `ArgumentParser` allows to do, but can potentially lead to unfixable problems.
- Argument parsing was a little bit handycraft work.

Meet [Plumbum](https://plumbum.readthedocs.io/), the swiss army knife for CLI packages:

- CLI has been refactored as a plumbum application.
- Tests for CLI parsing removed, since they were incompatible and we are using a well-tested CLI parsing library now after all.
- The application now includes 2 subcommands: `copy` and `update`.
- If called without args, or with 1, it will use `update`.
- If called with 2 args, it will use `copy` (this preserves old usage unchanged, unless you are copying from a template called `copy` or `update`; in such corner case, just call it as `copier copy copy destination` instead).
- `copier --help` and `copier --help-all` are more useful now.

And you might be wondering... Adding such a big dependency just for that? Well, actually plumbum supports more than just CLI applications: supports sane execution of local & remote commands, sane output coloring, sane user prompting... Possibly it will make Copier go to the next level as we keep on using it more. Just take a look at its docs! You'll love it :wink:

Of course, all that new CLI had to be reflected in API changes:

- `src_path` is now optional (`None` by default). In such case, copier will try to guess the `src_path` from the `.copier-answers.yml` file in the copy, or fail otherwise.
- `dst_path` is also now optional (`.` by default), meaning that without a `src_path`, you'll be trying to copy anything else to current `$PWD`.
- The builtin `_log` variable, which contains all necessary data for copy updates, now has also a `_src_path` key that points to the original source. This only happens when it was copied from a git repo or from an absolute path in the local disk, as relative paths are not easily reproducible in next updates (they depend on the user's CWD when running copier).
- A new `original_src_path` variable goes around, indicating the git repo or abs system path passed to `src_path` on the initial call. This is just to be stored in `_log` as explained.

```bash
copier gh:example/template destination
cd destination
echo my changes >> anywhere
git commit -m changing .
copier # Equals to `copier update .`
git commit -m 'copier update' .
```

:tada: You're updated!

This is an importan step to complete #88.

@Tecnativa TT20357

+++ copier/main.py
@@ -23,6 +23,7 @@ from .types import (
     AnyByStrDict,
     CheckPathFunc,
     OptBool,
+    OptStr,
     OptStrSeq,
     PathSeq,
     StrOrPath,
@@ -36,8 +37,8 @@ RE_TMPL = re.compile(r"\.tmpl$", re.IGNORECASE)
 
 
 def copy(
-    src_path: str,
-    dst_path: str,
+    src_path: OptStr = None,
+    dst_path: StrOrPath = ".",
     data: AnyByStrDict = None,
     *,
     exclude: OptStrSeq = None,
@@ -58,7 +59,8 @@ def copy(
     Arguments:
 
     - src_path (str):
-        Absolute path to the project skeleton. May be a version control system URL
+        Absolute path to the project skeleton. May be a version control system URL.
+        If `None`, it will be taken from `dst_path/.copier-answers.yml` or fail.
 
     - dst_path (str):
         Absolute path to where to render the skeleton
@@ -111,9 +113,13 @@ def copy(
         Remove the destination folder if the copy process or one of the tasks fail.
 
     """
-    repo = vcs.get_repo(src_path)
-    if repo:
-        src_path = vcs.clone(repo)
+    original_src_path = src_path
+    if src_path:
+        repo = vcs.get_repo(src_path)
+        if repo:
+            src_path = vcs.clone(repo)
+    elif src_path and not Path(src_path).is_absolute():
+        original_src_path = None
 
     conf, flags = make_config(**locals())
 
@@ -125,7 +131,7 @@ def copy(
             shutil.rmtree(dst_path, ignore_errors=True)
         raise
     finally:
-        if repo:
+        if repo and src_path:
             shutil.rmtree(src_path)
 
 
@@ -139,10 +145,11 @@ def copy_local(
     skip_if_exists: StrOrPathSeq,
     tasks: StrSeq,
     envops: Optional[AnyByStrDict],
+    original_src_path: str,
     flags: Flags,
 ) -> None:
 
-    render = get_jinja_renderer(src_path, data, extra_paths, envops)
+    render = get_jinja_renderer(src_path, data, extra_paths, envops, original_src_path)
 
     skip_if_exists = [render.string(pattern) for pattern in skip_if_exists]
     must_filter, must_skip = get_name_filters(exclude, include, skip_if_exists)
